requirements.txt
rembg
numpy==1.26
pillow
fastapi
uvicorn
